added attitude measurement
still need to test pid controller and quad copter controls

+++ Esc.cpp
@@ -7,12 +7,12 @@ Esc::Esc(int pin,long max,long min){
 	setup(pin,max,min);
 }
 
-void Esc::setup(int pin,long max,long min){
+void Esc::setup(int pin,short max,short min){
 	esc.attach(pin);
 	maxPulse = max;
 	minPulse = min;
 }
 
-void Esc::set(int width){
+void Esc::set(short width){
 	esc.writeMicroseconds(width);
 }
